Update Instruction.cs
This is a much cleaner way to achieve the same thing. My goal is for example to pair ax to rax and the assign a 16 bit mode to the instruction (since ax is 16 bits). That way I won't have to write a lot of spaghetti code, or as I like to call it SPAGHETT