Added MDUMP instruction
I debugged a bit since I found a critical bug and added the MDUMP instruction. This instruction dumps the register values as well as the memory and stack if the user wants to.

There are two modes, total (I do not recommend this for most applications) and incremental. Total dumps the whole memory and incremental dumps only the memory accessed by the program.